+++ mii/__init__.py
@@ -2,15 +2,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 # DeepSpeed Team
-try:
-    import grpc
-    from .pipeline import pipeline
-    from .server import serve
-    from .client import client
-except ImportError as e:
-    print("Warning: DeepSpeed-FastGen could not be imported:")
-    print(e)
-    pass
+import grpc
+from .api import client, serve, pipeline
 
 from .legacy import MIIServer, MIIClient, mii_query_handle, deploy, terminate, DeploymentType, TaskType, aml_output_path, MIIConfig, ModelConfig, get_supported_models
