+++ mii/launch/multi_gpu_server.py
@@ -10,7 +10,7 @@ import os
 from mii.config import ModelConfig
 from mii.grpc_related.modelresponse_server import serve_inference, serve_load_balancing
 from mii.grpc_related.restful_gateway import RestfulGatewayThread
-from mii.pipeline import async_pipeline
+from mii.api import async_pipeline
 
 
 def b64_encoded_config(config_str: str) -> ModelConfig:
